backend/requirements.txt
langchain>=0.1.0
langchain-openai==0.2.14
langchain-google-genai>=2.0.0
pinecone==7.3.0
openai==1.59.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.13.4
crewai==0.148.0
google-generativeai>=0.8.0
serpapi>=0.1.5
langchain-community>=0.0.20
tenacity>=8.4.0
loguru>=0.7.0
crewai_tools
pyyaml
# FastAPI and WebSocket dependencies
fastapi==0.118.0
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.19
# Vapi for voice support
vapi_server_sdk>=1.7.0
# Additional utilities
uuid-utils==0.9.0
